+++ src/FiscalYears/utils.js
@@ -16,7 +16,7 @@ export const useSaveFiscalYear = (mutator, onSave, method = 'POST') => {
         showCallout({
           messageId: 'ui-finance.fiscalYear.actions.save.success',
         });
-        onSave(savedFiscalYear.id);
+        onSave(savedFiscalYear);
 
         return savedFiscalYear;
       } catch (response) {