+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## (IN PROGRESS)
 
+* Close all current budgets when executing fiscal year rollover. Refs UIF-175
 * Set encumbrance detail for fiscal year rollover. Refs UIF-170
 * Set Fiscal year detail during fiscal year rollover. Refs UIF-254
 * Set Fund detail for fiscal year rollover. Refs UIF-169